python_face_service/requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
Pillow==10.0.1
numpy==1.24.3
dlib==19.24.2